% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lempelZivAl.R
\name{lempel.ziv}
\alias{lempel.ziv}
\title{Calcula la complejidad de Lempel Ziv normalizada, con el alfabeto que se desee}
\usage{
  lempel.ziv(s, alphabet = c(0, 1))
}
\arguments{
  \item{s}{vector de datos con la serie de tiempo a calcular}

  \item{alphabet}{alfabeto que se desea utilizar para el cálculo de la complejidad de Lempel Ziv}
}
\value{
  Complejidad de Lempel Ziv sin normalizar
}
\description{
  Función que a partir del alfabeto ingresado calcula la complejidad de Lempel Ziv
}
