% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lempelZiv.R
\name{lzc}
\alias{lzc}
\title{Calcula la complejidad de Lempel Ziv normalizada, con el alfabeto tradicional}
\usage{
  lzc(signal)
}
\arguments{
  \item{signal}{vector de datos con la serie de tiempo a calcular}
}
\value{
  Complejidad de Lempel Ziv normalizada
}
\description{
  Función que a partir del alfabeto tradicional (0,1), usando como umbral la media de la señal
}
