% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peEntropy.R
\name{pe}
\alias{pe}
\title{Calcula la entropía permutada}
\usage{
  pe(signal, embdim)
}
\arguments{
  \item{signal}{vector de datos con la serie de tiempo a calcular}

  \item{embdim}{embedding o tamaño de los patrones que se desean buscar en la serie de tiempo}
}
\value{
  Entropía permutada
}
\description{
  Función que a partir de los patrones ordinales que tiene la señal, calcula la entropía permutada
}
